generalize component test configuration #104
